backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
asyncpg
psycopg2-binary
h3
python-multipart
python-jose[cryptography]
passlib
python-dotenv
geojson
shapely
geoalchemy2
email-validator
pandas
geopandas
pyogrio
